+++ src/Local/Reconciliation/ResearchAutomation.Reconciliation.psm1
@@ -1,0 +1,3 @@
+Import-Module (Join-Path $PSScriptRoot 'ReconciliationEngine.psm1') -Force
+Import-Module (Join-Path $PSScriptRoot 'ReconciliationPersistence.psm1') -Force
+Export-ModuleMember -Function Test-RapIntegritySnapshot,Compare-RapIntegritySnapshot,New-RapReconciliationDecision,New-RapRecoveryPlan,Test-RapRecoveryReadBack,Invoke-RapRecoveryPlan,Initialize-RapReconciliationStore,New-RapReconciliationSqliteDependencies,Register-RapReconciliationFixture,Get-RapReconciliationSnapshot,Get-RapReconciliationAudit